+++ src/main/java/com/bigcart/user/managementservice/bigcartusermanagementservice/domain/service/GuestService.java
@@ -8,11 +8,11 @@ public interface GuestService {
 
     List<Guest> getAll();
 
-    Guest getById(int id);
+    Guest getById(long id);
 
     Guest add(Guest Guest);
 
-    Guest update(int id, Guest newGuest) throws IllegalAccessException;
+    Guest update(long id, Guest newGuest) throws IllegalAccessException;
 
-    boolean deleteById(int id);
+    boolean deleteById(long id);
 }
